Move `FolderDetails` to `:legacy:folder` module
