+++ legacy/core/src/main/java/com/fsck/k9/mailstore/ListenableMessageStore.kt
@@ -1,5 +1,6 @@
 package com.fsck.k9.mailstore
 
+import app.k9mail.legacy.folder.FolderDetails
 import app.k9mail.legacy.mailstore.CreateFolderInfo
 import com.fsck.k9.mail.FolderClass
 import java.util.concurrent.CopyOnWriteArraySet
