+++ legacy/core/src/main/java/com/fsck/k9/mailstore/MessageStore.kt
@@ -1,6 +1,7 @@
 package com.fsck.k9.mailstore
 
 import app.k9mail.legacy.account.Account.FolderMode
+import app.k9mail.legacy.folder.FolderDetails
 import app.k9mail.legacy.mailstore.CreateFolderInfo
 import app.k9mail.legacy.mailstore.FolderMapper
 import app.k9mail.legacy.mailstore.MessageMapper
