+++ legacy/folder/build.gradle.kts
@@ -5,3 +5,7 @@ plugins {
 android {
     namespace = "app.k9mail.legacy.folder"
 }
+
+dependencies {
+    implementation(projects.mail.common)
+}
