+++ legacy/folder/src/main/kotlin/app/k9mail/legacy/folder/FolderDetails.kt
@@ -1,0 +1,13 @@
+package app.k9mail.legacy.folder
+
+import com.fsck.k9.mail.FolderClass
+
+data class FolderDetails(
+    val folder: Folder,
+    val isInTopGroup: Boolean,
+    val isIntegrate: Boolean,
+    val syncClass: FolderClass,
+    val displayClass: FolderClass,
+    val notifyClass: FolderClass,
+    val pushClass: FolderClass,
+)
