+++ legacy/storage/src/main/java/com/fsck/k9/storage/messages/K9MessageStore.kt
@@ -1,6 +1,7 @@
 package com.fsck.k9.storage.messages
 
 import app.k9mail.legacy.account.Account.FolderMode
+import app.k9mail.legacy.folder.FolderDetails
 import app.k9mail.legacy.mailstore.CreateFolderInfo
 import app.k9mail.legacy.mailstore.FolderMapper
 import app.k9mail.legacy.mailstore.MessageMapper
@@ -11,7 +12,6 @@ import com.fsck.k9.mail.Flag
 import com.fsck.k9.mail.FolderClass
 import com.fsck.k9.mail.FolderType
 import com.fsck.k9.mail.Header
-import com.fsck.k9.mailstore.FolderDetails
 import com.fsck.k9.mailstore.LockableDatabase
 import com.fsck.k9.mailstore.MessageStore
 import com.fsck.k9.mailstore.StorageManager
