+++ legacy/storage/src/main/java/com/fsck/k9/storage/messages/UpdateFolderOperations.kt
@@ -1,10 +1,10 @@
 package com.fsck.k9.storage.messages
 
 import android.content.ContentValues
+import app.k9mail.legacy.folder.FolderDetails
 import app.k9mail.legacy.mailstore.MoreMessages
 import com.fsck.k9.mail.FolderClass
 import com.fsck.k9.mail.FolderType
-import com.fsck.k9.mailstore.FolderDetails
 import com.fsck.k9.mailstore.LockableDatabase
 import com.fsck.k9.mailstore.toDatabaseFolderType
