+++ legacy/storage/src/test/java/com/fsck/k9/storage/messages/UpdateFolderOperationsTest.kt
@@ -1,12 +1,12 @@
 package com.fsck.k9.storage.messages
 
 import app.k9mail.legacy.folder.Folder
+import app.k9mail.legacy.folder.FolderDetails
 import app.k9mail.legacy.folder.FolderType
 import app.k9mail.legacy.mailstore.MoreMessages
 import assertk.assertThat
 import assertk.assertions.isEqualTo
 import com.fsck.k9.mail.FolderClass
-import com.fsck.k9.mailstore.FolderDetails
 import com.fsck.k9.storage.RobolectricTest
 import org.junit.Test
 import com.fsck.k9.mail.FolderType as RemoteFolderType
