+++ legacy/ui/legacy/src/main/java/com/fsck/k9/ui/managefolders/FolderSettingsDataStore.kt
@@ -2,8 +2,8 @@ package com.fsck.k9.ui.managefolders
 
 import androidx.preference.PreferenceDataStore
 import app.k9mail.legacy.account.Account
+import app.k9mail.legacy.folder.FolderDetails
 import com.fsck.k9.mail.FolderClass
-import com.fsck.k9.mailstore.FolderDetails
 import com.fsck.k9.mailstore.FolderRepository
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
